configuration management

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,6 +1,6 @@
-#kill process killmenow
-
-exec { 'pkill':
-	command => 'pkill killmenow',
-	provider => 'shell',
+#kills a process
+exec {'kill_killmenow_proc':
+  command => 'pkill killmenow',
+  path    => '/usr/bin:/usr/sbin:/bin:/sbin',
+  onlyif  => 'pgrep killmenow',
 }
